+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Add block details ([\#35](https://github.com/forbole/big-dipper-2.0-solana/issues/35))
 - Add tx details ([\#36](https://github.com/forbole/big-dipper-2.0-solana/issues/36))
 - Add tokens list ([\#40](https://github.com/forbole/big-dipper-2.0-solana/issues/40))
+- Add tokens list searchbar ([\#45](https://github.com/forbole/big-dipper-2.0-solana/issues/45))